Simplify: merge @peartube/protocol into @peartube/host (#12)
Fold the thin protocol package into host, removing a package boundary and the
cross-package re-export/contract hop.

- Move create-client.js + event-map.js (+ types) into host/src/. create-client
  now imports the contract from ./contracts.js instead of @peartube/host.
  host/src/index.js re-exports contracts + event-map + create-client;
  host/package.json adds the ./events and ./create-client subpath exports and a
  @peartube/spec dependency.
- Repoint the 4 platform importers to @peartube/host (and @peartube/host/events),
  drop the @peartube/protocol dep from platform + app package.json, regenerate
  app/package-lock.json, and update root test/install:all scripts, the backend
  bundle manifest + working-tree freshness check, and the two regression tests
  that pinned the old protocol paths.
- Rewrite the living-doc architecture diagrams/tables to collapse the protocol
  layer into host (left docs/superpowers/* as historical record).
- Delete the @peartube/protocol package.

Validated locally (npm install --ignore-scripts): the moved create-client.test.mjs
passes against the new host barrel, all 4 CI-gated app regression tests pass
(incl. ci-lockfile-install-regression), zero @peartube/protocol references remain,
and the regenerated lockfile is valid.
