+++ QUICKSTART.md
@@ -63,7 +63,7 @@ Use focused package tests when possible:
 ```bash
 npm test --prefix packages/spec
 npm test --prefix packages/backend
-npm test --prefix packages/protocol
+npm test --prefix packages/host
 npm run desktop:native:test
 ```
