+++ package.json
@@ -35,12 +35,12 @@
     "build:eas:android": "npm run build:eas:android --prefix packages/app",
     "build:eas:all": "npm run build:eas:all --prefix packages/app",
     "typecheck": "npm run typecheck --prefix packages/platform",
-    "test": "npm test --prefix packages/spec && npm test --prefix packages/backend && npm test --prefix packages/host && npm test --prefix packages/protocol",
+    "test": "npm test --prefix packages/spec && npm test --prefix packages/backend && npm test --prefix packages/host",
     "test:watch": "npm run test:watch --prefix packages/backend",
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx --quiet",
     "lint:changed": "node scripts/lint-changed.mjs",
     "lint:fix": "npm run lint -- --fix",
-    "install:all": "npm install && npm install --prefix packages/spec && npm install --prefix packages/backend && npm install --prefix packages/host && npm install --prefix packages/protocol && npm install --prefix packages/platform && npm install --prefix packages/desktop-native && npm install --prefix packages/app --legacy-peer-deps"
+    "install:all": "npm install && npm install --prefix packages/spec && npm install --prefix packages/backend && npm install --prefix packages/host && npm install --prefix packages/platform && npm install --prefix packages/desktop-native && npm install --prefix packages/app --legacy-peer-deps"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^8.41.0",
