+++ packages/host/src/index.js
@@ -1,1 +1,3 @@
 export * from './contracts.js'
+export * from './event-map.js'
+export * from './create-client.js'
