--- packages/protocol/src/index.js
@@ -1,3 +1,0 @@
-export * from './event-map.js'
-export * from './create-client.js'
-export { HOST_ERROR_CODES, PROTOCOL_VERSION } from '@peartube/host'
